Update README.md
This project was made during the groupal project "Labs" module while coursing the Data Science bootcamp at soyHenry (www.soyhenry.com).

It consists in an ETL solution, for data architect and data engineer practice, the use of machine learning models for data scientist practice and a data analytics research, with insights, visualizations and dashboard.

We used Databricks as a integrated working platform, with AWS as a computing instances, storages and DB hosting provider.

Most of the code is written using Python, with libraries as Numpy, Pandas, XGB, SQLalchemy, skforecaster, folium, plotly and matplotlib. We used Spark, with part of the code written in Scala. The database was made using MySQL, and hosted in Amazon RDS service. The DataLake storage was provided by Amazon S3 and for the Data Warehouse we used Databrick's Delta Lake.

Even though our dataset is small, we decided to develop a scalable project, with clustering capabilities.

The project's notebooks are provided, as well as the dashboards and the project's report and documentation.

The project is in spanish language, as the bootcamp is located in Argentina, Latin America.